fix: Jest fetch configuration
In this project we are running Jest in jsdom environment (we are testing a web browser application).

JSDom does not implement the fetch web browser interface by default, which means we need to Polyfill it so we can run tests which perform calls to the method.

In this commit I am installing `jest-fetch-mock`; a library that makes it easy to mock HTTP requests.

In the future I would like to set up tests in a way they do perform actual HTTP requests rather than faking it, for a closer test experience for the end system, but for now this is a good solution to start writing tests.