feat: PDF document download logic.
The `WebPdfDocument` class extracts the `ArrayBuffer` from a PDF url by downloading it. It also ensures URLs point to valid PDF files. This makes it possible to them pass it down to the reMarkable client, which handles the buffer to upload the file.

On top of that, the Jest configuration has been changed. Now we use Node environment to skip the web-browser restrictions when downloading files, and `cross-fetch` so we use different HTTP clients in Node and Web Browser environments under the same `fetch` interface.