feat: `ConnectionManager`.
New service to pair our browser extension with the reMarkable Cloud and create sessions.

Currently the manager is a dummy class which calls the `RemarkableClient` authentication methods. I am planning to add some extra logic in future PRs to handle tokens stored in the extension, adding responsibilities enough to the service to justify its existence.