feat: Token storage with `ConnectionManager`.
`ConnectionManager` is now updated to persist tokens in a secured storage. The idea is to persist the device and session tokens so they are available whenever the user opens the web browser.

I still need to see how safe it is persisting the tokens in there, but at the moment this is a great approach to define the connection logic in isolation and integrate it easily in the service worker logic.