refactor: `ConnectionManager` & `UploadManager` extend `RemarkableManager`
All `RemarkableManagers` make use of the `ConfigurationManager` to communicate with the reMarkable Cloud using the user latest credential information, accessible through the `ConfigurationManager`.

Since all `RemarkableManager`s communicate with the reMarkable Cloud, this is a great way to keep a consistent way to exchange messages using the same user credentials among all services.