fix: fallback configuration for node dependencies in Storybook Webpack configuration.
It seems the node dependencies from the reMarkable library were causing problems with Storybook. Most likely it is due to the Device class being instantiated with an https client by default.

This should be fixed in the library, but in the meantime I am fixing here the Webpack configuration to ignore these node libraries when bundling the code.

First time a fix in Webpack works at first try, WOOOOW!