feat: Storage Manager
Adds a new `StorageManager` service that wraps the logic to persist data in the browser extension.

At the moment this is a wrapper for the `Storage` class, but in the future it will be use as a storage layer for abstracting the different types of storage available for browser extensions.