chore: `StorageManager` and `AuthenticationManager` now use always namespaces.
We force all storage to be allocated in a namespace, with a default namespace as fallback when it is not specified.

The `StorageManager` always sets a `Storage` with a namespace when instantiated. The `AuthenticationManager` uses a specific namespace where all device related information is stored. This namespace will be cleared when a new device is paired, securing all previos device data is erased, but allowing us to keep other information untouched (like web browser preferences for example).