feat: first successful attempt to create a content script UI.
I defined a simple button component that attaches to every `.pdf` link in the webpage. This is the idea behind the upload widget. Attaching the button automatically might not be the best approach in terms of usability, but for now we are going to leave it like that and then enhance the overall interaction.