feat: UploadWidget displayed as overlay component.
I want to display the upload widget when the user hovers the mouse over a link which can be uploaded to the reMarkable Cloud.

The UploadWidget component is now configured to handle the logic to be displayed in top of the element hovered.

Now I need to define the logic to set the overlay anchor in the DOM, so it is attached in the position I want to.