feat: Add Matcher for the Upload Widget.
Our Upload Widget is rendered withing a shadow root element which covers the screen. This element contains a wrapper element in which all Plasmo Content UI components are rendered, identified with a specific ID.

This matcher detects all elements within this container, and can be used to know when the user mouse is hovering our upload widget, keeping it visible in the UI until the mouse moves out of it and/or the element it was anchored.