feat: Mouse Tracker for upload widget.
The mouse tracker listens to events in the DOM viewport provided in the constructor and checks, for each DOM element being hovered, whether this contains a document that can be uploaded to the remarkable cloud.

If this is the case, the element is marked as an anchor with a specific ID, which can be then used by the Content Script UI as anchor reference to append the widget component next to the matching element.

With the new dependency we are able to write some unit tests that emulate the triggering of these events, providing us a good way to ensure the system behaves as we expect to.