chore: refactor logic to display upload button.
The upload button component was currently displayed very closed to the link it references.

With this minor changed we display the button at the end of the link, with some separation for a better UI display. We still need to extend the logic to ensure the button is never displayed out of the user view.