feat: PDF link anchor.
This method searches for potential `a` elements containing references to PDF documents, given an element which either is the link element or an element within it.

This logic will be used for attaching our upload widget next to PDF download links in the DOM. When the mouse passes over a link, this class will detect it as a potential anchor, which will then be used by the content script to attach itself to it.