feat: Anchor tracking manager.
Service object responsible of tracking mouse hover events along a specific viewport. On each new item hovered the object checks whether there is any potential anchor element to attach the upload widget component. If there is, labels the anchor candidate with a specific ID.

This will be used by the upload widget content script to detect when an element with this ID is introduced in the DOM, and attach the widget next to it.