feat: `getDocumentPreview` message for background script.
In order to get the full preview information of a web document (such as size), we need to perform certain HTTP requests which cannot be executed in the browser environment, but in the background worker environment.

With this message we allow elements from the web browser environment to request document previews by delegating the task to a background worker.