refactor: Message Manager.
I am struggling with a TS linter issue which only appears in Github CI. It seems to be related to the way I send messages to the background.

To keep the logic centered on a specific place I decided to define a MessageManager method which defines an interface with all the possible messages component can send to a background worker.

This reduces significantly the lint issues along the code, and make the whole logic way more maintainable. In future PRs I can extend it to do better error handling, something many parts of the extension are missing.