feat: Error serializer.
We need to define some logic to handle exceptions in the background worker.

If a message processing fails, the message should be answered with an error. The `ErrorSerializer` defines how these errors will be streamed through messages.

In a future PR all messages will handle exceptions returning error messages, and the message manager will make sure to handle these and return either the expected response or an Error back to the frontend.