feat: error handling in Background Worker.
The `Message` class is now extended to always capture any errors thrown during any message processing in a Background Worker. Catched errors are serialized and send back as a response to the message.

Since now all messages are performed through the `MessageManager`, this service has been updated to handle these exceptions, triggering a custom one with enough data for the front-end to know what went wrong.